# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hệ Thống Số Hoá Tài Liệu CTSV — REST API
  version: 1.0.0-draft
  description: |
    API RESTful cho hệ thống số hoá & quản lý tài liệu Công tác Sinh viên.
    File này là **contract chuẩn** (single source of truth).

    FastAPI-generated `openapi.json` chỉ là artifact để đối chiếu.
    **KHÔNG** được ghi đè file này từ implementation. Mọi thay đổi endpoint
    phải sửa YAML trước, rồi mới sửa code BE.

    Khi CI so sánh hai spec, dùng `oasdiff` (semantic diff) thay vì raw JSON
    diff — thứ tự field, metadata có thể khác nhưng semantic phải khớp.

    ## Quy ước

    - **Success**: envelope `{ success: true, data, total?, page?, limit? }`.
    - **Error**: RFC 7807 Problem Details (`application/problem+json`).
    - **Versioning**: URL prefix `/api/v1`.
    - **Auth**: `Authorization: Bearer <access_token>` cho mọi endpoint trừ login/refresh/health.
    - **Refresh token**: lưu ở `HttpOnly` cookie `rt`, KHÔNG trong response body.
    - **Idempotency**: header `Idempotency-Key: <uuid>` bắt buộc cho upload/trigger/activate.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.example.edu.vn/api/v1
    description: Production (placeholder)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: rt
      description: |
        HttpOnly cookie chứa refresh token. Set bởi `/auth/login`, clear bởi `/auth/logout`.
        Production phải có `Secure` + `SameSite=Lax` + `Path=/api/v1/auth`.
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ----- Domain enums -----
    DocumentScope:
      type: string
      enum: [PUBLIC, STUDENT_AFFAIRS, INTERNAL]
    DocumentStatus:
      type: string
      enum: [DRAFT, UNDER_REVIEW, APPROVED, ARCHIVED]
    JobStatus:
      type: string
      enum: [QUEUED, PROCESSING, SUCCEEDED, FAILED, CANCELLED]
    OCRReviewStatus:
      type: string
      enum: [PENDING, APPROVED, CORRECTED]
      description: |
        Trạng thái review cho từng block có `requires_review=true`.
        - PENDING: chờ staff/admin xem.
        - APPROVED: text OCR đã đúng, không cần sửa.
        - CORRECTED: text đã được sửa (cũ → mới lưu ở `original_text`).

    # ----- Domain entities -----
...
    Document:
      type: object
      required: [id, title, type, status, scope, latest_version, created_at, updated_at]
      properties:
        id: { type: string, example: "doc_01HXYZ..." }
        title: { type: string }
        type:
          type: string
          enum: [QUY_CHE, QUY_DINH, THONG_BAO, QUYET_DINH, HUONG_DAN, KHAC]
        status: { $ref: '#/components/schemas/DocumentStatus' }
        scope: { $ref: '#/components/schemas/DocumentScope' }
        code_number: { type: [string, "null"] }
        issuing_body: { type: [string, "null"] }
        effective_from: { type: [string, "null"], format: date }
        effective_to: { type: [string, "null"], format: date }
        latest_version: { type: integer }
        author_id: { type: string }
        tags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    UploadResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          required: [document_id, job_id, status]
          properties:
            document_id: { type: string }
            job_id: { type: string }
            status:
              $ref: '#/components/schemas/JobStatus'
...
    OCRBlock:
      type: object
      required: [id, version_id, page_number, block_index, text_content, confidence, bbox, requires_review, review_status, created_at, updated_at]
      properties:
        id: { type: string }
        version_id: { type: string }
        page_id: { type: [string, "null"] }
        page_number: { type: integer, minimum: 1 }
        block_index: { type: integer, minimum: 0 }
        text_content: { type: string }
        confidence:
          type: number
          minimum: 0
          maximum: 1
        bbox:
          type: array
          items: { type: number }
          minItems: 4
          maxItems: 4
          description: "[x_min, y_min, x_max, y_max] theo pixel ảnh render 300 DPI"
        requires_review: { type: boolean }
        review_status: { $ref: '#/components/schemas/OCRReviewStatus' }
        edited_text: { type: [string, "null"] }
        original_text: { type: [string, "null"] }
        job_id: { type: [string, "null"] }
        reviewed_by: { type: [string, "null"] }
        reviewed_at: { type: [string, "null"], format: date-time }
        processing_time_ms: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OCRPage:
      type: object
      required: [id, version_id, page_number, status, block_count, has_warnings, created_at]
      properties:
        id: { type: string }
        version_id: { type: string }
        page_number: { type: integer, minimum: 1 }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        image_key:
          type: [string, "null"]
          description: Object key nội bộ; chỉ dùng để quyết định có ảnh, không hiển thị cho client.
        status: { type: string }
        block_count: { type: integer, minimum: 0 }
        has_warnings: { type: boolean }
        created_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Chưa xác thực hoặc token hết hạn
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: Thiếu permission
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: Conflict (duplicate, version cũ)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    PayloadTooLarge:
      description: File vượt quá 50MB
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    UnsupportedMediaType:
      description: MIME không hỗ trợ
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
paths:
...
  /documents:
    get:
      tags: [documents]
      summary: Danh sách tài liệu (filter theo scope theo role)
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/DocumentStatus' }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string, description: "Full-text search trên title/code/tags" }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      tags: [documents]
      summary: Upload tài liệu mới (trả về job, OCR chạy async)
      description: |
        - Header `Idempotency-Key: <uuid>` BẮT BUỘC.
        - PDF â‰¤ 50MB; JPEG/PNG â‰¤ 10MB; MIME vÃ  magic bytes pháº£i khá»›p `application/pdf`, `image/jpeg` hoáº·c `image/png`.

        - Trả 202 Accepted + job_id (FE poll `/jobs/{id}`).
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title, type, scope]
              properties:
                file:
                  type: string
                  format: binary
                  description: "PDF â‰¤ 50MB; JPEG/PNG â‰¤ 10MB; MIME vÃ  magic bytes pháº£i khá»›p."

                title: { type: string }
                type:
                  type: string
                  enum: [QUY_CHE, QUY_DINH, THONG_BAO, QUYET_DINH, HUONG_DAN, KHAC]
                scope: { $ref: '#/components/schemas/DocumentScope' }
                issuing_body: { type: string }
                effective_from: { type: string, format: date }
                tags:
                  type: array
                  items: { type: string }
                change_summary:
                  type: string
                supersedes_document_id:
                  type: string
      responses:
        '202':
          description: Accepted — job đã được queue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /documents/{id}/versions/{vid}/ocr:
    get:
      tags: [ocr]
      summary: Lấy kết quả OCR (pages & blocks) của phiên bản tài liệu
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: vid
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: requires_review
          schema: { type: boolean }
        - in: query
          name: review_status
          schema: { $ref: '#/components/schemas/OCRReviewStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [version_id, ocr_status, requires_review, total_blocks, pending_reviews, pages, blocks]
                        properties:
                          version_id: { type: string }
                          ocr_status: { type: string }
                          requires_review: { type: boolean }
                          total_blocks: { type: integer }
                          pending_reviews: { type: integer }
                          pages:
                            type: array
                            items: { $ref: '#/components/schemas/OCRPage' }
                          blocks:
                            type: array
                            items: { $ref: '#/components/schemas/OCRBlock' }

        '401': { $ref: '#/components/responses/Unauthorized' }

        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /documents/{id}/versions/{vid}/ocr/pages/{page}/image:
    get:
      tags: [ocr]
      summary: Lấy ảnh PNG render 300 DPI của một trang OCR
      description: |
        Backend kiểm tra quyền đọc tài liệu và document scope trước khi lấy object
        `image_key` từ storage. Response là bytes PNG qua API; không trả public URL,
        presigned URL hoặc storage credential.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: vid
          required: true
          schema: { type: string }
        - in: path
          name: page
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: PNG trang OCR
          headers:
            Cache-Control:
              description: Không cho shared cache lưu ảnh tài liệu có kiểm soát scope.
              schema: { type: string, example: "private, no-store" }
          content:
            image/png:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /documents/{id}/versions/{vid}/ocr/blocks/{bid}:

    patch:
      tags: [ocr]
      summary: Duyệt hoặc chỉnh sửa một block OCR
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: vid
          required: true
          schema: { type: string }
        - in: path
          name: bid
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [review_status]
              properties:
                review_status: { $ref: '#/components/schemas/OCRReviewStatus' }
                text: { type: string }
      responses:
        '200':
          description: OK — trả về OCR block đã được cập nhật
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/OCRBlock' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /documents/{id}/versions/{vid}/ocr/batch-review:
    post:
      tags: [ocr]
      summary: Duyệt hàng loạt danh sách các block OCR hoặc chấp nhận toàn bộ block pending
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: vid
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accept_all_pending: { type: boolean, default: false }
                actions:
                  type: array
                  items:
                    type: object
                    required: [block_id, review_status]
                    properties:
                      block_id: { type: string }
                      review_status: { $ref: '#/components/schemas/OCRReviewStatus' }
                      text: { type: string }
      responses:
        '200':
          description: OK — trả về kết quả batch review
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [reviewed_count, remaining_pending_count, version_requires_review]
                        properties:
                          reviewed_count: { type: integer }
                          remaining_pending_count: { type: integer }
                          version_requires_review: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...